Demonstrate moving 3 tasks between cores
Very messy, but it proves that a thread can be interrupted,
context saved, and then context restored to a different core.

This example starts three tasks. They are rotated between three
different cores multiple times.

It seems that get_logical_core_id() is optimized by the compiler
to always return the value of the core the thread is initially run
on. Turning off optimizations makes it return the actual core.

Overly complicated synchronization within in ISR, but this is just
proof of concept without FreeRTOS yet.